Zadanie2/pyproject.toml
[tool.poetry]
name = "pwzn-pb-z1"
version = "0.1.0"
description = "Python w zastosowaniach naukowych - zadanie 2"
authors = ["Piotr Bienias"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.6.0"
Pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
